Disable unused features of chrono crate
Specifically, this disables "oldtime" feature, which pulls in "time"
crate for backwards-compatibility reasons. We don't use it, and it will
eventually be removed from chrono, so let's avoid depending on it by
mistake.

This does not remove the time crate from the dependency tree just yet,
as time crate is still used by hyper-old-types crate, but it prevents
Component Governance Alerts from misidentifying the dependency (thus
limiting the number of alerts).

Signed-off-by: Patryk Obara <dreamer.tan@gmail.com>